+++ components/brave_ads/core/internal/settings/settings.cc
@@ -5,8 +5,6 @@
 
 #include "brave/components/brave_ads/core/internal/settings/settings.h"
 
-#include <cstdint>
-
 #include "brave/components/brave_ads/common/notification_ad_feature.h"
 #include "brave/components/brave_ads/common/pref_names.h"
 #include "brave/components/brave_ads/core/internal/ads_client_helper.h"
@@ -14,14 +12,15 @@
 namespace brave_ads {
 
 int GetMaximumNotificationAdsPerHourSetting() {
-  int64_t ads_per_hour = AdsClientHelper::GetInstance()->GetInt64Pref(
-      prefs::kMaximumNotificationAdsPerHour);
+  int ads_per_hour =
+      static_cast<int>(AdsClientHelper::GetInstance()->GetInt64Pref(
+          prefs::kMaximumNotificationAdsPerHour));
 
   if (ads_per_hour == -1) {
-    ads_per_hour = static_cast<int64_t>(kDefaultNotificationAdsPerHour.Get());
+    ads_per_hour = kDefaultNotificationAdsPerHour.Get();
   }
 
-  return static_cast<int>(ads_per_hour);
+  return ads_per_hour;
 }
 
 }  // namespace brave_ads
